Basic addition of regex and sample file.
So I took a small bit of code which included comments from the sample
CMS-2 file we were sent. Wrote a little bit of code that searches the
file for lines that contain executable code and in-line comments. Prints
out results of the analysis.

This is just a small proof of concept for the regex idea. I will
continue working on this in the future to make it detect more patterns
to provide a more in-depth analysis of the code.